requirements.txt
Flask==2.3.3
gunicorn==21.2.0
aiogram==3.1.1
supabase==2.0.3
psycopg2-binary==2.9.7
redis==5.0.1
requests==2.31.0
aiohttp==3.8.6
python-dotenv==1.0.0
pydantic==2.3.0
python-dateutil==2.8.2
pytz==2023.3
APScheduler==3.10.4
structlog==23.1.0
cryptography==41.0.7